+++ packages/bundler/package.json
@@ -37,7 +37,7 @@
 		"style-loader": "^2.0.0",
 		"styled-components": "^5.2.1",
 		"url-loader": "^4.1.1",
-		"webpack": "5.20.1",
+		"webpack": "5.21.2",
 		"webpack-dev-middleware": "^3.7.2",
 		"webpack-hot-middleware": "^2.25.0"
 	},
