+++ packages/bundler/src/index.ts
@@ -3,14 +3,11 @@ import fs from 'fs';
 import getPort from 'get-port';
 import os from 'os';
 import path from 'path';
+import {Internals, WebpackOverrideFn} from 'remotion';
 import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
-import {
-	getOverrideFn,
-	overrideWebpackConfig,
-	WebpackOverrideFn,
-} from './override-webpack';
+import {overrideWebpackConfig} from './override-webpack';
 import {isUpdateAvailable} from './update-available';
 import {webpackConfig} from './webpack-config';
 
@@ -31,7 +28,8 @@ export const startServer = async (
 		userDefinedComponent,
 		outDir: tmpDir,
 		environment: 'development',
-		webpackOverride: options?.webpackOverride ?? getOverrideFn(),
+		webpackOverride:
+			options?.webpackOverride ?? Internals.getWebpackOverrideFn(),
 	});
 	const compiler = webpack(config);
