+++ packages/cli/package.json
@@ -45,7 +45,7 @@
 		"remotion": "^1.2.0",
 		"styled-components": "^5.1.1",
 		"typescript": "^4.1.2",
-		"webpack": "5.20.1",
+		"webpack": "5.21.2",
 		"xns": "^2.0.7"
 	},
 	"devDependencies": {
