+++ packages/core/package.json
@@ -32,6 +32,7 @@
 		"@types/react-dom": "^17.0.0",
 		"eslint": "^7.4.0",
 		"jest": "^26.6.3",
+		"webpack": "^5.21.2",
 		"prettier-plugin-organize-imports": "^1.1.1",
 		"ts-jest": "^26.4.4"
 	},
